Remove surface output

+++ src/gui/render_scalers.h
@@ -62,7 +62,7 @@ typedef ScalerLineHandler_t ScalerLineBlock_t[6][4];
 
 struct ScalerSimpleBlock_t {
 	const char* name         = {};
-	uint16_t gfxFlags        = 0;
+	uint8_t gfxFlags         = 0;
 	uint8_t xscale           = 0;
 	uint8_t yscale           = 0;
 	ScalerLineBlock_t Linear = {};
